wip: unit test to recreate issue #65, this will fail
this PR includes the reporoducer from @cirias on the issue.

note this does not include a fix, I looked at reverting the commit 1bb66b5 that this error started failing from but then other unit tests started failing, so figure someone that is familiar with the code is better off looking at a fix so that something else doesn't break.