feat(array): add enum support for array property types
I'm not super happy with digging down into t.Items.Enum directly, as it
leaves the door open for a panic if t.Items is nil. But as far as I can
tell, by the time t.arrayKeywords() is called, t.Items will always be
populated with a non-nil value.